pyproject.toml
[tool.poetry]
name = "django-docker"
version = "0.1.0"
description = ""
authors = ["Lee-Roy King <leeroywking@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
